chatlogic/requirements.txt
langchain==0.1.8
langchain-cli==0.0.21
langchain-openai==0.0.6
gql[all]==3.5.0
python-dotenv==1.0.1
httpx==0.26.0
openai==1.12.0